fix(core): Better handling and checks of user related flags
When sync_from_remote was called on a vine_data with a null user pointer,
the user flag would be (incorectly) set. This commit fixes this.

Repsectively in sync_to_remote, it was 'acceptable' to call sync_to_remote
on a buffer with the USER_SYNC set and with user==NULL.
This now triggers an assert.