perf(core): Minimize memset of vine_objects
Previously every vine_object would memset its allocation size.
This is problematic especialy for vine_data, that can range in up to gigabytes.
This commit minimizes the size of the memset to the metadata.
It also avoids redundantly memseting the base vine_object_s bytes,
as they are initialized by vine_object_register().